use runtime.reload() as a fallback for resetting icon
On browsers other than (currently) Firefox >= 59 it is not possible
to reset a tabId specific icon or badge by passing `null` for the icon
path and imageData / badge text.

To work around this, runtime.reload() is used upon user confirmation
when changing an setting where the code in question would have been
necessary.