remove support for Firefox 52-61
this guarantees that two api changes are available:
- the windowId parameter
- the ability to reset bages/icons

it also makes support for excluding hidden tabs considerably simpler

the tabs permission was only needed to check if the url had changed on a
tabs.onUpdated event. This was required from Firefox 59 onwards because
the tab-specific (not window-specific, though) icon/badge would be cleared
when a new page was loaded.