require font to be valid
It's considered to be invalid if it doesn't contain at least one
alphabetic character or apostrophe/quotation mark. This is a bit
simplified as e.g `1a` still causes problems but better than nothing.

If the input is considered it's ignored and the setting is reset to the
default value (sans-serif).